pyproject.toml
[tool.poetry]
name = "cognitive_apprenticeship"
version = "0.1.0"
description = ""
authors = ["Chris Proctor <chris@chrisproctor.net>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
django-tinymce = "^3.3.0"
django-crispy-forms = "^1.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
